Refactor a bunch of code
* This is so that we can accomodate the mellanox switch, so this commit
breaks down the "run_command" function into three functions, one to login;
another to run the commands, and then one to logout.

* Update the regex for the brocade switch to accurately match the password prompt.
The previous string would sometimes match the warning message the said
"some password has not changed" and that would cause us to hang in the while loop.

* Add the set of commands for the mellanox switch.

* Also add a bunch of debugging statements that I will remove when it's ready
to deploy.